Move getToolstack() into LocalApplication and adapt tests
